refactor: 페이지 타입분리
