+++ src/components/UI/MeetingTime.tsx
@@ -1,11 +1,5 @@
 import React, { JSX } from 'react'
-
-interface props {
-  time: string
-  onClick?: () => void // 시간 선택 핸들러
-  isSelected?: boolean
-  disabled: boolean
-}
+import { MeetingTimeProps } from '../../types/meeting'
 
 // 면접 시간 선택 버튼 컴포넌트
 const MeetingTime = ({
@@ -13,7 +7,7 @@ const MeetingTime = ({
   onClick,
   isSelected,
   disabled,
-}: props): JSX.Element => {
+}: MeetingTimeProps): JSX.Element => {
   return (
     <button
       className={`font-pretendardBold text-white rounded-3xl text-center border-2 border-mainColor w-[80px] h-[40px] ${isSelected ? 'bg-mainColor' : 'bg-none'} ${
